TVG-108 Move fetching a TVGuide to the TVGuide component
